KrakenTypes are all exported and their own package
