+++ src/README.md
@@ -1,0 +1,9 @@
+~/go     $GOPATH
+~/go/bin $GOBIN
+~/go/pkg/-.....
+~/go/src/github.com/CryptoTrader/main.go
+~/go/src/github.com/KrakenTypes/KrakenTypes.go
+
+(in KrakenTypes)
+go build
+go install
